# DomainIntelWeb/openapi.json
{
  "components": {
    "schemas": {
      "ArchiveState": {
        "properties": {
          "archived_to": {
            "title": "Archived To",
            "type": "string"
          }
        },
        "required": [
          "archived_to"
        ],
        "title": "ArchiveState",
        "type": "object"
      },
...
      "CountState": {
        "properties": {
          "deleted": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "boolean"
              }
            ],
            "title": "Deleted"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "CountState",
        "type": "object"
      },
...
      "DailyIdentity": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          }
        },
        "required": [
          "date",
          "category",
          "key"
        ],
        "title": "DailyIdentity",
        "type": "object"
      },
      "DailyState": {
        "properties": {
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "dates": {
            "items": {
              "type": "string"
            },
            "title": "Dates",
            "type": "array"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "origins": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Origins",
            "type": "object"
          },
          "selection_scope": {
            "const": "current_page",
            "title": "Selection Scope",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "selection_scope",
          "dates",
          "counts",
          "origins"
        ],
        "title": "DailyState",
        "type": "object"
      },
      "DeleteDailyRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DailyIdentity"
            },
            "maxItems": 5000,
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "DeleteDailyRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IndustryCreate": {
        "properties": {
          "folder": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Folder",
            "type": "string"
          },
          "name": {
            "default": "",
            "maxLength": 120,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "folder"
        ],
        "title": "IndustryCreate",
        "type": "object"
      },
      "IndustryMutationState": {
        "properties": {
          "folder": {
            "title": "Folder",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "folder",
          "name"
        ],
        "title": "IndustryMutationState",
        "type": "object"
      },
      "IndustryRename": {
        "properties": {
          "folder": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Folder",
            "type": "string"
          },
          "name": {
            "default": "",
            "maxLength": 120,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "folder"
        ],
        "title": "IndustryRename",
        "type": "object"
      },
      "IndustryState": {
        "additionalProperties": true,
        "properties": {
          "folder": {
            "title": "Folder",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "periodic_enabled": {
            "default": false,
            "title": "Periodic Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "folder",
          "name"
        ],
        "title": "IndustryState",
        "type": "object"
      },
...
      "OverviewState": {
        "properties": {
          "chain": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Chain",
            "type": "array"
          },
          "entities": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Entities",
            "type": "array"
          },
          "industry": {
            "additionalProperties": true,
            "title": "Industry",
            "type": "object"
          },
          "latest_document_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Document Date"
          },
          "source_categories": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Source Categories",
            "type": "object"
          },
          "stats": {
            "additionalProperties": true,
            "title": "Stats",
            "type": "object"
          }
        },
        "required": [
          "industry",
          "stats",
          "chain",
          "entities",
          "source_categories"
        ],
        "title": "OverviewState",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "IntDog Local API",
    "version": "4.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/industries": {
      "get": {
        "operationId": "industries_api_industries_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IndustryState"
                  },
                  "title": "Response Industries Api Industries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Industries",
        "tags": [
          "industries"
        ]
      },
      "post": {
        "operationId": "create_industry_api_industries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndustryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndustryMutationState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Industry",
        "tags": [
          "industries"
        ]
      }
    },
    "/api/industries/{folder}": {
      "delete": {
        "operationId": "archive_industry_api_industries__folder__delete",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Industry",
        "tags": [
          "industries"
        ]
      },
      "patch": {
        "operationId": "rename_industry_api_industries__folder__patch",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndustryRename"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndustryMutationState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Industry",
        "tags": [
          "industries"
        ]
      }
    },
...
    "/api/industries/{folder}/daily": {
      "delete": {
        "operationId": "delete_daily_api_industries__folder__daily_delete",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteDailyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Daily",
        "tags": [
          "daily"
        ]
      },
      "get": {
        "operationId": "daily_api_industries__folder__daily_get",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "title",
              "enum": [
                "title",
                "category",
                "source"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 500,
              "title": "Cursor",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Daily",
        "tags": [
          "daily"
        ]
      }
    },
...
    "/api/industries/{folder}/overview": {
      "get": {
        "operationId": "overview_api_industries__folder__overview_get",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overview",
        "tags": [
          "industries"
        ]
      }
    },
...
  }
}